ingesta_datos/requirements.txt
asttokens==2.4.1
azure-core==1.30.1
azure-functions==1.18.0
azure-storage-blob==12.19.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
idna==3.6
ipykernel==6.29.4
ipython==8.22.2
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
mutagen==1.47.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pycryptodomex==3.20.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
youtube-transcript-api==0.6.2
yt-dlp==2024.3.10
